ai/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.6
pymupdf==1.25.3
pytesseract==0.3.13
Pillow==11.1.0
sentence-transformers==3.4.1
spacy==3.8.4
transformers==4.49.0
python-multipart==0.0.20
